Aiogram/requirements.txt

aiogram==3.21.0
aiohttp==3.12.15
PySocks==1.7.1
Telethon==1.40.0
